+++ src/main/java/nextstep/ladder/domain/ladder/Ladder.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain.ladder;
 
 import java.util.Objects;
