+++ src/main/java/nextstep/ladder/domain/ladder/strategy/RandomStrategy.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain.ladder.strategy;
 
 import java.util.ArrayList;
 import java.util.List;
