+++ src/main/java/nextstep/ladder/domain/ladder/strategy/Strategy.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain.ladder.strategy;
 
 import java.util.List;
