+++ src/main/java/nextstep/ladder/domain/participant/Participant.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain.participant;
 
 import java.util.Objects;
