+++ src/test/java/nextstep/ladder/domain/ladder/LadderTest.java
@@ -1,5 +1,6 @@
-package nextstep.ladder;
+package nextstep.ladder.domain.ladder;
 
+import nextstep.ladder.domain.ladder.Ladder;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
