+++ src/test/java/nextstep/ladder/domain/ladder/LineTest.java
@@ -1,5 +1,6 @@
-package nextstep.ladder;
+package nextstep.ladder.domain.ladder;
 
+import nextstep.ladder.domain.ladder.Line;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
